src/lcd_interface/requirements.txt
RPi.GPIO>=0.7.0
paho-mqtt>=1.6.1
